+++ starsky/starskytest/starsky.foundation.webtelemetry/Helpers/FlushOnApplicationStoppingTest.cs
@@ -3,7 +3,10 @@ using Microsoft.ApplicationInsights;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using starsky.foundation.platform.Models;
 using starsky.foundation.webtelemetry.Helpers;
+using starskytest.FakeMocks;
+using starskytest.starsky.foundation.platform.Services;
 
 namespace starskytest.starsky.foundation.webtelemetry.Helpers
 {
@@ -64,6 +67,47 @@ namespace starskytest.starsky.foundation.webtelemetry.Helpers
 			await new FlushApplicationInsights(serviceProvider).FlushAsync();
 			var telemetryClient = serviceProvider.GetService<TelemetryClient>();
 			Assert.IsNotNull(telemetryClient);
-		} 
+		}
+
+
+		[TestMethod]
+		public void GetTelemetryClientTestKeyShouldHitLogger()
+		{
+			var logger = new FakeIWebLogger();
+			new FlushApplicationInsights(new ServiceCollection()
+				.BuildServiceProvider(), new AppSettings{ApplicationInsightsInstrumentationKey = "t"}, logger).GetTelemetryClient();
+			Assert.AreEqual(1, logger.TrackedInformation.Count);
+			Assert.AreEqual("TelemetryClient is null on exit", logger.TrackedInformation[0].Item2);
+		}
+		
+		[TestMethod]
+		public void FlushApplicationInsights_HitLogger_Sync()
+		{
+			var logger = new FakeIWebLogger();
+			new FlushApplicationInsights(new ServiceCollection()
+				.BuildServiceProvider(), new AppSettings{ApplicationInsightsInstrumentationKey = "t"}, logger).Flush();
+			Assert.AreEqual(1, logger.TrackedInformation.Count);
+			Assert.AreEqual("TelemetryClient is null on exit", logger.TrackedInformation[0].Item2);
+		}
+		
+		[TestMethod]
+		public async Task FlushApplicationInsights_HitLogger_aSync()
+		{
+			var logger = new FakeIWebLogger();
+			await new FlushApplicationInsights(new ServiceCollection()
+				.BuildServiceProvider(), new AppSettings{ApplicationInsightsInstrumentationKey = "t"}, logger).FlushAsync();
+			Assert.AreEqual(1, logger.TrackedInformation.Count);
+			Assert.AreEqual("TelemetryClient is null on exit", logger.TrackedInformation[0].Item2);
+		}
+
+		
+		[TestMethod]
+		public void GetTelemetryClientTestIgnore()
+		{
+			var logger = new FakeIWebLogger();
+			new FlushApplicationInsights(new ServiceCollection()
+				.BuildServiceProvider(), new AppSettings{ApplicationInsightsInstrumentationKey = ""}, logger).GetTelemetryClient();
+			Assert.AreEqual(0, logger.TrackedInformation.Count);
+		}
 	}
 }